requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==7.3.0
cryptography==41.0.3
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4

